% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{my_template}
\alias{my_template}
\title{Template function to send an email with 'blastula' ensuring the formatting
is done properly (full width and with limited padding all round the body
of the message)}
\usage{
my_template(
  html_body = NULL,
  html_header = NULL,
  html_footer = NULL,
  title = NULL,
  content_width = "100000px",
  font_family = "Helvetica, sans-serif"
)
}
\description{
Template function to send an email with 'blastula' ensuring the formatting
is done properly (full width and with limited padding all round the body
of the message)
}
\author{
Gianluca Baio
}
\keyword{Tutees}
\keyword{emails}
