% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{Sends an email using a combination of 'blastula' and 'Microsoft365R'}
\usage{
send_email(
  email,
  subject,
  from = "stats.tutor@ucl.ac.uk",
  to,
  cc = NULL,
  bcc = NULL,
  attachment = NULL,
  ...
)
}
\arguments{
\item{email}{An email created using 'blastula', typically the output of a
call to the function \code{write_email}}

\item{subject}{A string of text with the subject in the email}

\item{from}{A string with the email address from which the message should
be sent. Deafaults to the 'Stats Tutor' shared mailbox
("stats.tutor@ucl.ac.uk"). It can be also specified to be the
account whose credentials have been authenticated by 'Microsoft365', or
another account that has right to send from this authenticated user.}

\item{to}{A string with the email address of the recipient(s) of the email.
Multiple addresses can be included in a vector}

\item{cc}{A string with the email address of the CC recipient(s) of the
email. Multiple addresses can be included in a vector}

\item{bcc}{A string with the email address of the BCC recipient(s) of the
email. Multiple addresses can be included in a vector}

\item{attachment}{The full path to an attachment file}

\item{\dots}{Additional options.}
}
\description{
Sends an email using a combination of 'blastula' and 'Microsoft365R'
}
\examples{
\dontrun{
email=write_email(
  system.file("templates","example.qmd",package="tutees")
)
send_email(
  email=email,
  subject="Just a quick example...",
  to="g.baio@ucl.ac.uk",
  cc=c("g.baio@ucl.ac.uk","gianluca@statistica.it")
)
}
}
\author{
Gianluca Baio
}
\keyword{Tutees}
\keyword{emails}
