% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{write_email}
\alias{write_email}
\title{Generic function to write the body of an email message, starting from a
template provided in the file 'file'.}
\usage{
write_email(file)
}
\arguments{
\item{file}{A string with the path pointing to the text file that contains
the text of the email to be sent. The file is parsed through \code{blastula},
which uses the function \code{md} (based on \code{glue}) to "glue" the text
into a single string to be parsed in html code. For this reason, \code{R} code
\strong{must} be included in the syntax \code{\{CODE HERE\}}, for example
\code{\{data$name[i]\}} and \code{Rmarkdown}. See the template in the
example.
"chunks" don't work.}
}
\description{
Generic function to write the body of an email message, starting from a
template provided in the file 'file'.
}
\examples{
\dontrun{
email=write_email(
  system.file("templates","example.qmd",package="tutees")
)
send_email(
  email=email,
  subject="Just a quick example...",
  to="g.baio@ucl.ac.uk"
)
}
}
\author{
Gianluca Baio
}
\keyword{Tutees}
\keyword{emails}
